This improves the performance by allowing v8 to optimize based on parameters. Still not quite as fast as lodash.sumby identity though.

Current master:
```
lodash.sumby identity x 25,055,277 ops/sec ±1.60% (82 runs sampled)
sum-by identity x 14,032,427 ops/sec ±4.37% (80 runs sampled)
lodash.sumby iteratee x 6,492,915 ops/sec ±2.04% (85 runs sampled)
sum-by iteratee x 6,544,549 ops/sec ±1.87% (85 runs sampled)
```

This PR:
```
lodash.sumby identity x 24,768,103 ops/sec ±0.90% (85 runs sampled)
sum-by identity x 19,912,574 ops/sec ±1.93% (87 runs sampled)
lodash.sumby iteratee x 5,277,562 ops/sec ±6.78% (73 runs sampled)
sum-by iteratee x 12,844,107 ops/sec ±1.29% (88 runs sampled)
```