@queicherius 2018-10-08
Very ,very nice stuff. I did not even know that V8 deoptimized functions without parameters so heavily.